+++ packages/component-news/src/components/CardGridNews/index.js
@@ -24,13 +24,14 @@ const gridRow = (feed, cardButton) => (
       eventFormat="inline"
       eventLocation={feed.location}
       clickable={!!feed.buttonLink}
+      cardLink={feed.alias}
       clickHref={feed.buttonLink}
       title={feed.title}
       body={`<p class="card-text text-dark">${feed.content}</p>`}
       image={feed.imageUrl}
       imageAltText={feed.title}
       linkLabel={feed.eventButtonText}
-      linkUrl={feed.eventButtonUrl}
+      linkUrl={feed.eventButtonUrl || feed.buttonLink}
       buttons={[
         {
           ariaLabel: cardButton.text,